Move `Cors` into `arg_enums` module to the rest of CLI data structures
